HARP-7872: Move zoom and ppi to MapOperators.ts
