+++ @here/harp-datasource-protocol/lib/ExprEvaluator.ts
@@ -31,6 +31,7 @@ import { ComparisonOperators } from "./operators/ComparisonOperators";
 import { FeatureOperators } from "./operators/FeatureOperators";
 import { FlowOperators } from "./operators/FlowOperators";
 import { InterpolationOperators } from "./operators/InterpolationOperators";
+import { MapOperators } from "./operators/MapOperators";
 import { MathOperators } from "./operators/MathOperators";
 import { MiscOperators } from "./operators/MiscOperators";
 import { ObjectOperators } from "./operators/ObjectOperators";
@@ -229,3 +230,4 @@ ExprEvaluator.defineOperators(ArrayOperators);
 ExprEvaluator.defineOperators(InterpolationOperators);
 ExprEvaluator.defineOperators(ObjectOperators);
 ExprEvaluator.defineOperators(FeatureOperators);
+ExprEvaluator.defineOperators(MapOperators);
